requirements.txt
python-docx==0.8.11
Flask==2.3.2
pywebview==4.0.1
pyinstaller==5.10.0